Move alias to internals
